rename Amodels to Models and other minor changes
